requirements.txt
absl-py
adal
alabaster
anyio
argon2-cffi
argon2-cffi-bindings
arrow
artifacts-keyring
asttokens
astunparse
async-generator
async-lru
async-timeout
atomicwrites
attrs
autopep8
Babel
backcall
beautifulsoup4
black
bleach
cachetools
clang
click
cloudpickle
comm
contourpy
curio
cycler
debugpy
decorator
defusedxml
docrepr
docutils
entrypoints
exceptiongroup
executing
fastjsonschema
filelock
flake8
flatbuffers
fonttools
fqdn
fsspec
gast
google-auth
google-auth-oauthlib
google-pasta
graphviz
grpcio
h11
h5py
httpcore
httpx
imagesize
importlib_metadata
iniconfig
ipykernel
ipyparallel
ipython
ipython-genutils
ipywidgets
isoduration
jaraco.classes
jaraco.context
jaraco.functools
jedi
Jinja2
joblib
json5
jsonpointer
jsonschema
jsonschema-specifications
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets
keras
keyring
kiwisolver
libclang
lightgbm
llvmlite
Markdown
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mccabe
mdurl
mistune
ml-dtypes
mmdnn
more-itertools
mpmath
mypy-extensions
namex
nbclassic
nbclient
nbconvert
nbformat
nest-asyncio
networkx
notebook
notebook_shim
numba
numpy
oauthlib
opencv-python
opt-einsum
optree
outcome
overrides
pandas
pandocfilters
parso
pathspec
patsy
pickleshare
pillow
platformdirs
pluggy
prometheus_client
prompt_toolkit
protobuf
psutil
pure_eval
py
pyasn1
pyasn1_modules
pycodestyle
pydot
pydotplus
pyflakes
Pygments
PyJWT
pyparsing
pyrsistent
pytest
pytest-asyncio
python-dateutil
python-json-logger
pytz
pywin32
pywin32-ctypes
pywinpty
PyYAML
pyzmq
qtconsole
QtPy
redis
referencing
requests-oauthlib
rfc3339-validator
rfc3986-validator
rich
rpds-py
rsa
scikit-learn
scipy
Send2Trash
setuptools
shap
six
slicer
sniffio
snowballstemmer
sortedcontainers
soupsieve
Sphinx
sphinx-rtd-theme
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jquery
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
stack-data
statsmodels
sympy
tensorboard
tensorboard-data-server
tensorflow
tensorflow-estimator
tensorflow-intel
termcolor
terminado
testpath
threadpoolctl
tinycss2
toml
tomli
torch
torchvision
tornado
traitlets
trio
types-python-dateutil
typing_extensions
tzdata
uri-template
wcwidth
webcolors
webencodings
websocket-client
Werkzeug
wheel
widgetsnbextension
wincertstore
wrapt
yapf
zipp
zstandard